reflect updates to contract state
When sending transactions updated contract state is updated via a call
back that is called 24 times.  Is this ganache or default ethreum
behavior?

this should be watching for an event that is fired when value is updated
in smart contract.

added spinner for wait state but not wired into any functionality yet.